Added CLI functionality (help, passable arguments)
