+++ setup/create-wp-project/tests/e2e/create-wp-project.test.js
@@ -1,17 +1,23 @@
 const { exec } = require('promisify-child-process');
 const timeout = 180000;
+const del = require('del'); // eslint-disable-line
 
-beforeEach(() => {
+const testingFolder = 'temp';
+
+beforeEach(async() => {
+  await del(`./${testingFolder}`);
 });
 
-afterEach(async () => {
-  // cleanup
+afterEach(async() => {
+  await del(`./${testingFolder}`);
 });
-
 
 // Test readdirAsync
-test('that create-wp-project script completes successfully', async() => {
-  const { stdout, stderr } = await exec('npx create-wp-project');
+test('test create-wp-project script (local version)', async() => {
+  const { stdout, stderr } = await exec(
+    `cd ${testingFolder} &&
+    node ~/my-packages/eightshift-frontend-libs/setup/create-wp-project/create-wp-project.js --projectName="Test Project" --url="asdasd.local" --description="This is a description" --noSummary`
+  );
   console.log('stdout', stdout);
   console.log('stderr', stderr);
 }, timeout);
